Fix example folder structure
The Arduino IDE requires sketches to be in a folder of the same name.

By placing the examples folder inside the library folder, the example will be accessible via File > Examples > ConcurrentArduino > Blink.

The Arduino Library Specification states that the examples folder should be named with all lower case letters.